+++ src/include/tyranscript/tyran_rb_tree.h
@@ -20,7 +20,7 @@ typedef struct stree_node{
 typedef struct sroot{
   struct stree_node* root;
   void* (*key)(struct stree_node* node);
-  int64_t (*compare)(void* keyA, void* keyB);
+  int (*compare)(void* keyA, void* keyB);
 }tree_root;
 
 typedef struct siterator{
@@ -29,7 +29,7 @@ typedef struct siterator{
 
 extern tree_root* new_simple_rbtree();
 extern tree_root* new_rbtree(void* (*key_function_pointer)(struct stree_node* node),
-			   int64_t (*compare_function_pointer)(void* keyA, void* keyB));
+			   int (*compare_function_pointer)(void* keyA, void* keyB));
 extern void* rb_tree_insert(tree_root* root, void* node);
 extern void* rb_tree_delete(tree_root* root, void* key);
 extern void* search_rbtree(tree_root root, void* key);
